server/requirements.txt
diffusers==0.2.4
# transformers scipy
ftfy
flask_cors